[Projects] TaskApp global component added - Along with TaskInline & KanbanTask
